fix(ci): create branch before creating pull request (#206)
