--- TEST.SH.md
@@ -1,2 +1,0 @@
-mv test.sh "$(echo TEST.sh | tr [:lower:] [:upper:]).md" | true
-mv test.sh "$(echo TEST.sh | tr [:lower:] [:upper:]).md" | true
