+++ content/aips/4B301050BD35-AIP-ADD-STMATIC-TO-AAVE-V3-POLYGON.md
