+++ package.json
@@ -9,11 +9,15 @@
     "David Truong <david@aave.com>"
   ],
   "dependencies": {
+    "@types/bs58": "^4.0.1",
+    "@types/node-fetch": "^2.6.2",
     "ajv": "^8.11.0",
-    "bs58": "^4.0.1",
+    "bs58": "^5.0.0",
+    "dotenv": "^16.0.3",
     "ethers": "^5.7.0",
+    "ipfs-only-hash": "^4.0.0",
     "markdown-to-json": "^0.5.4",
-    "node-fetch": "^3.2.10",
+    "node-fetch": "^2.6.7",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
   },
@@ -33,7 +37,7 @@
   },
   "scripts": {
     "parse-markdown": "m2j ./content/aips/*.md -c -o ./content/ipfs-aips/all-aips.json",
-    "upload:pinata-thegraph": "node scripts/aip-uploader.js && rm tmp",
+    "upload:pinata-thegraph": "ts-node scripts/aip-uploader.ts && rm tmp",
     "validation:aip-schema": "ts-node scripts/aip-schema-validation.ts",
     "ci:generate-readme": "ts-node scripts/generate-readme.ts"
   }
